+++ netharn/util/util_misc.py
@@ -2,6 +2,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 import numpy as np
 import ubelt as ub
+import re
 
 
 class SupressPrint():
@@ -84,3 +85,24 @@ class FlatIndexer(ub.NiceRepr):
         """
         base = self.cums[outer] - self.lens[outer]
         return base + inner
+
+
+def strip_ansi(text):
+    r"""
+    Removes all ansi directives from the string.
+
+    References:
+        http://stackoverflow.com/questions/14693701/remove-ansi
+        https://stackoverflow.com/questions/13506033/filtering-out-ansi-escape-sequences
+
+    Examples:
+        >>> line = '\t\u001b[0;35mBlabla\u001b[0m     \u001b[0;36m172.18.0.2\u001b[0m'
+        >>> escaped_line = strip_ansi(line)
+        >>> assert escaped_line == '\tBlabla     172.18.0.2'
+    """
+    # ansi_escape1 = re.compile(r'\x1b[^m]*m')
+    # text = ansi_escape1.sub('', text)
+    # ansi_escape2 = re.compile(r'\x1b\[([0-9,A-Z]{1,2}(;[0-9]{1,2})?(;[0-9]{3})?)?[m|K]?')
+    ansi_escape3 = re.compile(r'(\x9B|\x1B\[)[0-?]*[ -/]*[@-~]', flags=re.IGNORECASE)
+    text = ansi_escape3.sub('', text)
+    return text